Refactoring code after feedback
