+++ cg/services/illumina/backup/backup_service.py
@@ -1,5 +1,4 @@
 import subprocess
-from datetime import datetime
 from pathlib import Path
 
 from cg.apps.slurm.slurm_api import SlurmAPI
@@ -19,18 +18,13 @@ from cg.meta.tar.tar import TarAPI
 from cg.models.cg_config import PDCArchivingDirectory
 from cg.models.run_devices.illumina_run_directory_data import IlluminaRunDirectoryData
 from cg.services.illumina.backup.encrypt_service import IlluminaRunEncryptionService
-from cg.services.illumina.backup.models import PdcEncryptionKey, PdcSequencingFile
 from cg.services.illumina.backup.utils import (
-    DsmcOutput,
-    get_latest_pdc_archived_sequencing_run,
-    get_latest_pdc_encryption_key,
-    is_pdc_encryption_key,
-    is_pdc_sequencing_path,
+    get_latest_archived_encryption_key_path,
+    get_latest_archived_sequencing_run_path,
 )
 from cg.services.pdc_service.pdc_service import PdcService
 from cg.store.models import IlluminaSequencingRun
 from cg.store.store import Store
-from cg.utils.date import convert_string_to_datetime_object
 from cg.utils.time import get_elapsed_time, get_start_time
 
 
@@ -104,8 +98,8 @@ class IlluminaBackupService:
             sequencing_run.device.internal_id
         )
 
-        archived_key: Path = self.get_latest_archived_encryption_key_path(dsmc_output=dsmc_output)
-        archived_run: Path = self.get_latest_archived_sequencing_run_path(dsmc_output=dsmc_output)
+        archived_key: Path = get_latest_archived_encryption_key_path(dsmc_output=dsmc_output)
+        archived_run: Path = get_latest_archived_sequencing_run_path(dsmc_output=dsmc_output)
 
         if not self.dry_run:
             return self._process_run(
@@ -285,77 +279,6 @@ class IlluminaBackupService:
         self.pdc.retrieve_file_from_pdc(
             file_path=str(archived_file), target_path=str(retrieved_file)
         )
-
-    @staticmethod
-    def parse_dsmc_output_sequencing_path(dsmc_output: list[str]) -> list[PdcSequencingFile]:
-        """Parses the DSMC command output to extract validated sequencing paths."""
-        validated_responses = []
-        for line in dsmc_output:
-            if is_pdc_sequencing_path(line):
-                parts = line.split()
-
-                fileDateTime: datetime = convert_string_to_datetime_object(
-                    f"{parts[DsmcOutput.DATE_COLUMN_INDEX]} {parts[DsmcOutput.TIME_COLUMN_INDEX]}"
-                )
-
-                query_response: PdcSequencingFile = PdcSequencingFile(
-                    dateTime=fileDateTime,
-                    path=Path(parts[DsmcOutput.PATH_COLUMN_INDEX]),
-                )
-                validated_responses.append(query_response)
-
-        return validated_responses
-
-    @classmethod
-    def get_latest_archived_sequencing_run_path(cls, dsmc_output: list[str]) -> Path | None:
-        """Get the path of the archived sequencing run from a PDC query."""
-        validated_sequencing_paths: list[PdcSequencingFile] = cls.parse_dsmc_output_sequencing_path(
-            dsmc_output
-        )
-
-        archived_run: PdcSequencingFile | None = get_latest_pdc_archived_sequencing_run(
-            validated_sequencing_paths
-        )
-
-        if archived_run:
-            LOG.info(f"Sequencing run found: {archived_run}")
-            return archived_run.path
-
-    @staticmethod
-    def parse_dsmc_output_key_path(dsmc_output: list[str]) -> list[PdcEncryptionKey]:
-        """Parses the DSMC command output to extract validated encryption keys."""
-        validated_responses = []
-        for line in dsmc_output:
-            if is_pdc_encryption_key(line):
-                parts: list[str] = line.split()
-
-                fileDateTime: datetime = convert_string_to_datetime_object(
-                    f"{parts[DsmcOutput.DATE_COLUMN_INDEX]} {parts[DsmcOutput.TIME_COLUMN_INDEX]}"
-                )
-
-                query_response: PdcEncryptionKey = PdcEncryptionKey(
-                    dateTime=fileDateTime,
-                    path=Path(parts[DsmcOutput.PATH_COLUMN_INDEX]),
-                )
-
-                validated_responses.append(query_response)
-
-        return validated_responses
-
-    @classmethod
-    def get_latest_archived_encryption_key_path(cls, dsmc_output: list[str]) -> Path | None:
-        """Get the encryption key for the archived sequencing run from a PDC query."""
-        validated_encryption_keys: list[PdcEncryptionKey] = cls.parse_dsmc_output_key_path(
-            dsmc_output
-        )
-
-        archived_encryption_key: PdcEncryptionKey | None = get_latest_pdc_encryption_key(
-            validated_encryption_keys
-        )
-
-        if archived_encryption_key:
-            LOG.info(f"Encryption key found: {archived_encryption_key}")
-            return archived_encryption_key.path
 
     def validate_is_run_backup_possible(
         self,
